+++ src/lib/Person.java
@@ -1,0 +1,11 @@
+package lib;
+
+public class Person{
+
+	private String employeeId;
+	private String firstName;
+	private String lastName;
+	private String idNumber;
+	private String address;
+
+}
